back/requirements.txt
flask>=1.0.1
eventlet>=0.24.0
torch==1.6.0
numpy>=1.14.5
cn2an==0.5.2
ruamel.yaml==0.15.87
torchmetrics==0.7.2
flask-socketio==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0